procs: Move "view" implementation into a separate file
This is consistent with other procs.
